Extracted ReadBundler class to khmer/utils.py
